im an idiot, modelsApi
