+++ lib/api/modelsApi.ts
@@ -1,10 +1,28 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
-import {Core, Endpoints}  from '@/conf/cfg';
-interface Model {
-  id: string;
-  // other fields...
+import { Endpoints }  from '@/conf/cfg';
+interface UnitCosts {
+  input?: string;
+  output?: string;
+  image?: number;
 }
 
+interface Model {
+  id: string;
+  object: string;
+  proxy_to?: string;
+  owned_by?: string;
+  limit?: string;
+  unit_costs?: UnitCosts;
+  allowed_for: string[];
+  max_images?: number;
+  multiple_of?: number;
+  unit_cost?: string;
+}
+
+interface Response {
+  data: Model[];
+  object: string;
+}
 export const modelsApi = createApi({
   reducerPath: 'modelsApi',
   baseQuery: fetchBaseQuery({ baseUrl: Endpoints.NAGA_BASE_URL }),
