+++ lib/store.ts
@@ -1,18 +1,26 @@
 import {configureStore} from '@reduxjs/toolkit';
 import screenSizeReducer from './features/screenSize/screenSizeSlice';
 import menuOpenReducer from './features/isMenuOpen/isMenuOpenSlice';
-import modelsSliceReducer from './features/models/modelsSlice';
 import limitsSliceReducer from './features/limits/limitsSlice';
-export const makeStore = () => {
+import { modelsApi } from './api/modelsApi';
+import { setupListeners } from '@reduxjs/toolkit/query';
 
-return configureStore({
-        reducer: {
-            screenSize: screenSizeReducer,
-            isMenuOpen: menuOpenReducer,
-            models: modelsSliceReducer,
-            limits: limitsSliceReducer
-        }
-    })
+export const makeStore = () => {
+  const store = configureStore({
+    reducer: {
+      screenSize: screenSizeReducer,
+      isMenuOpen: menuOpenReducer,
+      [modelsApi.reducerPath]: modelsApi.reducer,
+      limits: limitsSliceReducer
+    },
+    middleware: (getDefaultMiddleware) =>
+      getDefaultMiddleware().concat(modelsApi.middleware),
+  });
+
+  // Set up listeners for RTK Query
+  setupListeners(store.dispatch);
+
+  return store;
 }
 
 export type AppStore = ReturnType<typeof makeStore>
